+++ tickets_api/tickets_api/serve.py
@@ -336,7 +336,7 @@ QUERY_EVENT_SUBQUERY = {
                'ILIKE %s',
         'description': 'Value for recipient_group tag'
                        'as set by the rule expert.',
-        'label': 'Recipient Group Tag',
+        'label': 'Recipient Group',
         'exp_type': 'string',
     },
     'recipient_group_icontains': {
@@ -345,7 +345,7 @@ QUERY_EVENT_SUBQUERY = {
                'ILIKE concat(\'%%\', %s, \'%%\')',
         'description': 'Value for recipient_group tag'
                        'as set by the rule expert - substring match.',
-        'label': 'Recipient Group Tag contains',
+        'label': 'Recipient Group contains',
         'exp_type': 'string',
     },
     'recipient-address_is': {